Clean up most of the comment parsing
